cfon_invoice_bulk_ocr_import_v2_src/requirements.txt
pypdf>=4.2.0
pdfplumber>=0.11.4
pillow>=10.4.0
pytesseract>=0.3.13
python-dateutil>=2.9.0
